requirements.txt
sentence-transformers
torch
pinecone
openai
python-dotenv
langchain
langchain-community
langchain-google-genai
google-genai
pandas
numpy
fastapi
uvicorn
streamlit
langchain-core
openpyxl
langchain-pinecone
gspread